Relocate visual effects in project structure (#89)
* [BED-105] Relocate visuals to feedback package

* Rewrite server_id selection in personal dev deployment workflow (#87)

* Convert names correctly

* Revert "Synchronize PR revert"

* Emphasize important message parts

* Add missing arguments to message

* Force min 1 block build height

* Associate lootspawner to team

* Remove unused field

* Debug teams to ensure proper configuration

* Enhance message

* Add debug serialization

* Refresh the team

* Refactor code

* Introduce dedicated debug stage

* Debug entire setup

* Remove builder style

* Refactor team handling

* Refactor team getter

* Add debug messages

* Remove explicit setting

* Bring back builder style

* Reconfigure teams

* Remove Debug logging

* Clear list when (re)configuring to avoid duplications

* Revert changes

* Adjust messages

* Reconfigure teams

* Refactor codebase

* Debug teams before every stage

* Switch to named args

* Refactor readability

* Debug configured teams

* Remove debug

* Switch to JacksonLocation

* Rollback

* Fix invalid data type

* Debug current team

* Reset team reference

* Debug team

* Debug phase & size

* Adjust early return

* Introduce configured field

* Adjust configuration

* Change condition

* Remove  stage logic

* Debug phase and size

* Start next phase

* Pre increment phase

* Inline feedback message

* Refactor stage behavior

* Adjust condition

* Rollback to code from branch BED-54

* Fix success message

* Fix success message

* Remove explicit refreshing team reference

* Debug teams

* Refactor stage handling

* Remove debugging

* Remove builder pattern

* Refactor stage handling

* Refactor stage handling

* Rename fields

* Switch to named args

* Rewrite bed configuration

* Remove builder pattern

* Log location

* Use cloned objects

* Clone objects before using them

* Debug location

* Apply offset to locations

* Fix non-configured spawner

* Fix invalid saved location

* Cleanup codebase

* Remove DebugStage

* Add documentation

* Finish setup

* Refactor location types

* Refactor structure

* Remove obsolete class

* Rewrite GameMap class

* Write GameMap into file

* Write files into maps folder

* Run setup completion

* Refactor welcome/completion stage

* Introduce MapSetupManager singleton

* Add missing messages

* Add documentation

* Bump version to 1.0.0-alpha.3

* Fix invalid team color name

* Annotate team list

* Fix typo

* Introduce visualization

* Add support for Display Entities

* Remove support for Display Entities

* Introduce BlockVisualizer

* Rewrite visualization

* Debug rendering

* Remove debugging

* Debug color

* Remove debugging

* Switch to RGB colors

* Adjust inheritance hierarchy

* Introduce fake block renderings

* Introduce fake block renderings

* Fix invalid location correction

* Correct entity position

* Remove position correction

* Bump version to 1.0.0-alpha.4

* Abstract block visualisation

* Refactor name

* Refactor block list type

* Introduce MultiVerticalBlockRenderer

* Refactor names

* Change color

* Render bounding-box with team colors

* Introduce sight line

* Change particle type

* Fix player tracking bug

* Track eye location only once

* Fix invalid direction

* Render visuals when unsneaking

* Adjust eye location

* Lower eye height

* Refactor common rendering

* Render lobby spawnpoint

* Render bed bounding boxes

* Refactor rendered block type

* Render fake bed

* Render bed in team color

* render bed at top half

* Fix typo in 'visualization'

* Introduce height renderings

* Adjust config values

* Increase particle count

* Add more particles

* Add more particles

* Adjust spawn distance

* Introduce utility classes

* Refactor fake block visualization

Took 43 seconds

* Center entity

Took 6 minutes

* Adjust location

Took 3 minutes

* Print location as debug

Took 5 minutes

* Adjust rotations

Took 8 minutes

* Fix incorrect bed locations

Took 5 minutes

* Refactor correction

Took 4 minutes

* Correct location for other blocks

Took 7 minutes

* Make chest look at player

Took 3 minutes

* Remove unclear corrections for chests

Took 4 minutes

* Correct chest locations

Took 5 minutes

* Adjust correction

Took 5 minutes

* [BED-105] Relocate visuals to feedback package

* [BED-107] Fix invalid environment

Took 10 minutes

* [BED-107] Use JSON entries for server selection

Took 6 minutes

* Revert "Rewrite server_id selection in personal dev deployment workflow (#87)" (#88)

This reverts commit e9f2a7d34675f0d0f158c4aaaada656d979effc0.

* Rewrite deployments

Took 6 minutes
