+++ .github/workflows/deploy-to-personal-environment.yml
@@ -11,6 +11,7 @@ on:
 env:
   host_url: ${{ secrets.PTERODACTYL_HOST }}
   api_key: ${{ secrets.PTERODACTYL_API_KEY }}
+  server_ids: ${{ secrets.PTERODACTYL_SERVER_IDS }}
   vip_person: waddlespam
 
 jobs:
@@ -18,16 +19,15 @@ jobs:
     if: "!startsWith(github.event.head_commit.message, 'Merge')" # Only deploy if the push is not triggered by a merge action
     environment: local development
     runs-on: ubuntu-latest
+    outputs:
+      server_id: ${{ steps.set-id.outputs.SERVER_ID }}
     steps:
       - name: Evaluate target server
+        id: set-id
         run: |
-          if [[ "${{ github.actor }}" == "${{ env.vip_person }}" ]]; then
-            echo "server_id=${{ secrets.PTERODACTYL_SERVER_ID_2 }}" >> $GITHUB_ENV
-            echo "Using environment of vip person ${{ env.vip_person }}"
-          else
-            echo "server_id=${{ secrets.PTERODACTYL_SERVER_ID }}" >> $GITHUB_ENV
-            echo "Using default environment"
-          fi
+          SERVER_ID=$(echo '${{ env.server_ids }}' | jq -r '."${{ env.vip_person }}"')
+          echo "SERVER_ID=$SERVER_ID"
+          echo "SERVER_ID=$SERVER_ID" >> "$GITHUB_OUTPUT"
 
   stop_server:
     environment: local development
@@ -37,7 +37,7 @@ jobs:
     steps:
       - name: Stop local server
         run: |
-          curl "${{ env.host_url }}/api/client/servers/${{ env.server_id }}/power" \
+          curl "${{ env.host_url }}/api/client/servers/${{ needs.evaluate_target.outputs.server_id }}/power" \
           -H 'Accept: application/json' \
           -H 'Content-Type: application/json' \
           -H 'Authorization: Bearer ${{ env.api_key }}' \
@@ -48,6 +48,7 @@ jobs:
     environment: local development
     runs-on: ubuntu-latest
     needs:
+      - evaluate_target
       - stop_server
     steps:
       - uses: actions/checkout@v4
@@ -70,7 +71,7 @@ jobs:
         with:
           panel-host: ${{ env.host_url }}
           api-key: ${{ env.api_key }}
-          server-id: ${{ env.server_id }}
+          server-id: ${{ needs.evaluate_target.outputs.server_id }}
           source: build/libs/*-deploy.jar
           target: /plugins/
           restart: true
@@ -79,11 +80,12 @@ jobs:
     environment: local development
     runs-on: ubuntu-latest
     needs:
+      - evaluate_target
       - deploy
     steps:
       - name: Start local server
         run: |
-          curl "${{ env.host_url }}/api/client/servers/${{ env.server_id }}/power" \
+          curl "${{ env.host_url }}/api/client/servers/${{ needs.evaluate_target.outputs.server_id }}/power" \
           -H 'Accept: application/json' \
           -H 'Content-Type: application/json' \
           -H 'Authorization: Bearer ${{ env.api_key }}' \
