+++ .github/workflows/deploy-to-staging-emvironment.yml
@@ -14,7 +14,7 @@ env:
 jobs:
   stop_server:
     if: "startsWith(github.event.head_commit.message, 'Merge')" # Only deploy if the push is not triggered by a merge action
-    environment: local development
+    environment: staging
     runs-on: ubuntu-latest
     steps:
       - name: Stop local server
@@ -27,7 +27,7 @@ jobs:
           -d '{"signal":"stop"}'
 
   deploy:
-    environment: local development
+    environment: staging
     runs-on: ubuntu-latest
     needs:
       - stop_server
@@ -58,7 +58,7 @@ jobs:
           restart: true
 
   start_server:
-    environment: local development
+    environment: staging
     runs-on: ubuntu-latest
     needs:
       - deploy
