+++ src/main/java/net/alphalightning/bedwars/feedback/visual/Visualization.java
@@ -1,4 +1,4 @@
-package net.alphalightning.bedwars.setup.visual;
+package net.alphalightning.bedwars.feedback.visual;
 
 import org.jetbrains.annotations.NotNull;
