+++ src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/MinBuildHeightConfigurationStage.java
@@ -7,8 +7,8 @@ import net.alphalightning.bedwars.setup.map.GameMapSetup;
 import net.alphalightning.bedwars.setup.map.MapSetup;
 import net.alphalightning.bedwars.setup.map.stages.HeightConfiguration;
 import net.alphalightning.bedwars.setup.map.stages.Stage;
-import net.alphalightning.bedwars.setup.visual.impl.HeightRenderer;
-import net.alphalightning.bedwars.setup.visual.impl.HeightVisualization;
+import net.alphalightning.bedwars.feedback.visual.impl.HeightRenderer;
+import net.alphalightning.bedwars.feedback.visual.impl.HeightVisualization;
 import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
