+++ src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/TeamChestConfigurationStage.java
@@ -8,10 +8,10 @@ import net.alphalightning.bedwars.setup.map.jackson.Team;
 import net.alphalightning.bedwars.setup.map.stages.LocationConfiguration;
 import net.alphalightning.bedwars.setup.map.stages.Stage;
 import net.alphalightning.bedwars.setup.map.stages.TeamConfiguration;
-import net.alphalightning.bedwars.setup.visual.impl.BlockEdgeRenderer;
-import net.alphalightning.bedwars.setup.visual.impl.BlockEdgeVisualization;
-import net.alphalightning.bedwars.setup.visual.impl.FakeBlockRenderer;
-import net.alphalightning.bedwars.setup.visual.impl.FakeBlockVisualization;
+import net.alphalightning.bedwars.feedback.visual.impl.BlockEdgeRenderer;
+import net.alphalightning.bedwars.feedback.visual.impl.BlockEdgeVisualization;
+import net.alphalightning.bedwars.feedback.visual.impl.FakeBlockRenderer;
+import net.alphalightning.bedwars.feedback.visual.impl.FakeBlockVisualization;
 import net.alphalightning.bedwars.translation.NamedTranslationArgument;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TranslatableComponent;
