+++ src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/TeamSpawnpointConfigurationStage.java
@@ -8,10 +8,10 @@ import net.alphalightning.bedwars.setup.map.jackson.Team;
 import net.alphalightning.bedwars.setup.map.stages.LocationConfiguration;
 import net.alphalightning.bedwars.setup.map.stages.Stage;
 import net.alphalightning.bedwars.setup.map.stages.TeamConfiguration;
-import net.alphalightning.bedwars.setup.visual.impl.MultiBlockRenderer;
-import net.alphalightning.bedwars.setup.visual.impl.MultiBlockVisualization;
-import net.alphalightning.bedwars.setup.visual.impl.SingleLineRenderer;
-import net.alphalightning.bedwars.setup.visual.impl.SingleLineVisualization;
+import net.alphalightning.bedwars.feedback.visual.impl.MultiBlockRenderer;
+import net.alphalightning.bedwars.feedback.visual.impl.MultiBlockVisualization;
+import net.alphalightning.bedwars.feedback.visual.impl.SingleLineRenderer;
+import net.alphalightning.bedwars.feedback.visual.impl.SingleLineVisualization;
 import net.alphalightning.bedwars.translation.NamedTranslationArgument;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TranslatableComponent;
